pyproject.toml
[project]
name = "image-prompt-library"
version = "0.1.0"
description = "Local-first image prompt reference library"
requires-python = ">=3.9"
dependencies = [
  "fastapi>=0.110",
  "uvicorn>=0.27",
  "pydantic>=2",
  "python-multipart>=0.0.9",
  "pillow>=10",
  "httpx>=0.27",
]

[project.optional-dependencies]
dev = ["pytest>=8"]

[tool.setuptools]
packages = ["backend", "backend.routers", "backend.services"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
